Main/med-report-assistant/requirements.txt
langchain
langchain-openai
langgraph
streamlit
pypdf
python-dotenv
faiss-cpu
langchain-community

